requirements.txt
activations==0.1.0
datasets==2.0.0
file_utils==0.0.1
flax==0.4.1
jax==0.3.6
jaxlib==0.3.5
marisa_trie==0.7.7
numpy==1.20.1
pandas==1.2.4
parso==0.7.0
pytorch_lightning==1.6.1
PyYAML==6.0
regex==2021.4.4
responses==0.20.0
scikit_learn==1.0.2
tensorflow==2.8.0
tokenization==1.0.7
tokenizers==0.12.1
torch==1.11.0
tqdm==4.59.0
transformers==4.18.0
